+++ NEWS.md
@@ -1,3 +1,10 @@
+# stuff for next release
+
+* `compare_models` is deprecated in favor of `loo_compare`
+* The `kfold` method now has a `cores` argument and parallelizes by fold rather
+than by Markov chain (unless otherwise specified), which should be much more
+efficient when many cores are available.
+
 # rstanarm 2.18.1
 
 ### Bug fixes